rename util -> pkg
