+++ cmd/tink-cli/cmd/hardware/id.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/tinkerbell/tink/util"
+	"github.com/tinkerbell/tink/pkg"
 
 	"github.com/spf13/cobra"
 	"github.com/tinkerbell/tink/client"
@@ -29,7 +29,7 @@ var idCmd = &cobra.Command{
 			if err != nil {
 				log.Fatal(err)
 			}
-			b, err := json.Marshal(util.HardwareWrapper{Hardware: hw})
+			b, err := json.Marshal(pkg.HardwareWrapper{Hardware: hw})
 			if err != nil {
 				log.Fatal(err)
 			}
