+++ pkg/hardware_wrapper.go
@@ -1,4 +1,4 @@
-package util
+package pkg
 
 import (
 	"encoding/json"
